requirements.txt
scipy==1.5.2
pillow===10.4.0
pandas==1.1.3
torchvision==0.12.0
kornia==0.6.1
regex==2024.11.6
ftfy==6.2.3
numpy==1.24.4
opencv-python==4.5.4.60
timm==1.0.15
tqdm==4.66.2
scikit-learn==1.3.2
torch==1.11.0
transformers==4.35.2
open-clip-torch==2.23.0
matplotlib==3.7.5
